// src/shared/api/generated/model/productListItemResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CoreBank API Specification
 * Corebank REST API 명세서
 * OpenAPI spec version: v1.0.0
 */
import type { ProductListItemResponseProductGroup } from './productListItemResponseProductGroup';

export interface ProductListItemResponse {
  productId?: number;
  productCode?: string;
  productName?: string;
  productGroup?: ProductListItemResponseProductGroup;
  summary?: string;
  baseRate?: number;
  maxRate?: number;
  minTermMonths?: number;
  maxTermMonths?: number;
  minAmount?: number;
  maxAmount?: number;
  newProduct?: boolean;
}