// src/shared/api/generated/model/productTermsViewResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CoreBank API Specification
 * Corebank REST API 명세서
 * OpenAPI spec version: v1.0.0
 */

export interface ProductTermsViewResponse {
  termsId?: number;
  termsName?: string;
  version?: string;
  required?: boolean;
  viewRequired?: boolean;
  content?: string;
  viewedAt?: string;
  viewExpiresAt?: string;
}