// src/shared/api/generated/model/termsItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CoreBank API Specification
 * Corebank REST API 명세서
 * OpenAPI spec version: v1.0.0
 */

export interface TermsItem {
  termsId?: number;
  termsName?: string;
  version?: string;
  required?: boolean;
  viewRequired?: boolean;
  displayOrder?: number;
}